album のコンテンツメニューをAPIで読み込むようにした

+++ fuel/app/modules/album/classes/controller/api.php
@@ -49,6 +49,68 @@ class Controller_Api extends \Controller_Site_Api
 	}
 
 	/**
+	 * Api get_menu
+	 * 
+	 * @access  public
+	 * @return  Response
+	 */
+	public function get_menu($id = null)
+	{
+		$response = '';
+		try
+		{
+			$this->check_response_format('html');
+
+			$is_detail  = (bool)\Input::get('is_detail', 0);
+			$id = (int)$id;
+			$album = Model_Album::check_authority($id);
+			$this->check_browse_authority($album->public_flag, $album->member_id);
+
+			$menus = array();
+			if ($album->member_id == $this->u->id)
+			{
+				if (!Site_Util::check_album_disabled_to_update($album->foreign_table, true))
+				{
+					if (!$is_detail) $menus[] = array('tag' => 'divider');
+					$menus[] = array('href' => 'album/edit/'.$id, 'icon_term' => 'form.do_edit');
+					$menus[] = array('icon_term' => 'form.do_delete', 'attr' => array(
+						'class' => $is_detail ? 'js-simplePost' : 'js-ajax-delete',
+						'data-uri' => $is_detail ? 'album/delete/'.$album->id : 'album/api/delete/'.$id.'.json',
+						'data-msg' => term('form.delete').'します。よろしいですか。',
+						'data-parent' => 'main_item_'.$id,
+					));
+				}
+			}
+			elseif (is_enabled('notice'))
+			{
+				$is_watched = \Notice\Model_MemberWatchContent::get_one4foreign_data_and_member_id('album_', $id, $this->u->id);
+				$menus[] = array('icon_term' => $is_watched ? 'form.do_unwatch' : 'form.do_watch', 'attr' => array(
+					'class' => 'js-update_toggle',
+					'data-uri' => 'member/notice/api/update_watch_status/album/'.$id,
+					'data-msg' => $is_watched ? term('form.watch').'を解除しますか？' : term('form.watch').'しますか？',
+				));
+			}
+
+			$response = \View::forge('_parts/dropdown_menu', array('menus' => $menus));
+			$status_code = 200;
+		}
+		catch(\HttpNotFoundException $e)
+		{
+			$status_code = 404;
+		}
+		catch(\HttpForbiddenException $e)
+		{
+			$status_code = 403;
+		}
+		catch(\FuelException $e)
+		{
+			$status_code = 500;
+		}
+
+		$this->response($response, $status_code);
+	}
+
+	/**
 	 * Api albums
 	 * 
 	 * @access  public
@@ -72,7 +134,8 @@ class Controller_Api extends \Controller_Site_Api
 	 */
 	public function post_delete($id = null)
 	{
-		$response = array('status' => 0);
+		$response = array('status' => 0, 'error_messages' => array());
+		$response['error_messages']['default'] = sprintf('%sの%sに失敗しました。', term('album'), term('form.delete'));
 		try
 		{
 			\Util_security::check_csrf();
@@ -80,22 +143,48 @@ class Controller_Api extends \Controller_Site_Api
 			$id = (int)$id;
 			if (\Input::post('id')) $id = (int)\Input::post('id');
 
-			\DB::start_transaction();
 			$album = Model_Album::check_authority($id, $this->u->id, 'member');
+			if ($result = Site_Util::check_album_disabled_to_update($album->foreign_table))
+			{
+				throw new \DisableToUpdateException($result['message']);
+			}
+			\DB::start_transaction();
 			$album->delete();
 			\DB::commit_transaction();
 
 			$response['status'] = 1;
+			$response['message'] = sprintf('%sを%sしました。', term('album'), term('form.delete'));
 			$status_code = 200;
 		}
 		catch(\HttpNotFoundException $e)
 		{
 			$status_code = 404;
 		}
+		catch(\HttpForbiddenException $e)
+		{
+			$status_code = 403;
+		}
+		catch(\DisableToUpdateException $e)
+		{
+			$response['error_messages']['absolute'] = $e->getMessage();
+			$status_code = 403;
+		}
+		catch(\HttpInvalidInputException $e)
+		{
+			$status_code = 400;
+		}
+		catch(\Database_Exception $e)
+		{
+			$status_code = 500;
+		}
 		catch(\FuelException $e)
 		{
+			$status_code = 500;
+		}
+		if ($status_code == 500)
+		{
 			if (\DB::in_transaction()) \DB::rollback_transaction();
-			$status_code = 400;
+			$response['error_messages']['500'] = $response['error_messages']['default'];
 		}
 
 		$this->response($response, $status_code);
@@ -122,7 +211,7 @@ class Controller_Api extends \Controller_Site_Api
 			$album = Model_Album::check_authority($id, $this->u->id, 'member');
 			if ($result = Site_Util::check_album_disabled_to_update($album->foreign_table))
 			{
-				throw new \DisableToUpdatePublicFlagException($result['message']);
+				throw new \DisableToUpdateException($result['message']);
 			}
 			list($public_flag, $model) = \Site_Util::validate_params_for_update_public_flag($album->public_flag);
 
@@ -144,7 +233,7 @@ class Controller_Api extends \Controller_Site_Api
 
 			return \Response::forge($response, $status_code);
 		}
-		catch(\DisableToUpdatePublicFlagException $e)
+		catch(\DisableToUpdateException $e)
 		{
 			$response = json_encode(array('error' => array('message' => $e->getMessage())));
 			$status_code = 403;
